+++ packages/styleguide/src/components/Preview/Preview.docs.mdx
@@ -1,7 +1,7 @@
 import { md } from './../../utils/';
 import { Preview, ComponentDocs, CodeJSX } from './../';
 import PreviewForDocs from './Preview';
-import InteractForDOcs from './Interact';
+import InteractForDOcs from './Interact/Interact';
 import { Bar, BarItem } from './../Bar';
 
 # Preview
